server: move ConnTestSuite to conn_stmt_test.go
